requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Faker==4.0.3
gitdb==4.0.5
GitPython==3.1.2
html5lib==1.0.1
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
lxml==4.5.0
mosspy==1.0.8
mossum @ git+https://github.com/hjalti/mossum@2a22ef644c7f87e391b501023cd8b79d875a5eec
numpy==1.18.4
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.14.0
smmap==3.0.4
soupsieve==2.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
xlwt==1.3.0
